Modulo de Domicilio y asignar docimicilio a una direccion

+++ local/database/migrations/2022_04_21_020319_add_municipality_to_users_table.php
@@ -25,8 +25,12 @@ class AddMunicipalityToUsersTable extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('municipality');
-        });
+        if (Schema::hasColumn('users', 'municipality'))
+        {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('municipality');
+            });
+        }
+        
     }
 }
